requirements.txt
openpyxl==3.1.5
Pillow==11.1.0
pymongo==4.11.1
tkcalendar==1.6.1
